requirements.txt
# AWS SDK
boto3==1.34.34
botocore==1.34.34

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==12.1.1

# Vector DB & Search
opensearch-py==2.4.2
requests-aws4auth==1.2.3

# Web Framework
Flask==3.0.2
Flask-Cors==4.0.0
gunicorn==22.0.0

# Text Processing & Embeddings
langchain==0.3.27
langchain-community==0.3.27
sentence-transformers==2.3.1
tiktoken==0.6.0

# Environment & Configuration
python-dotenv==1.0.1
pydantic==2.12.5
pydantic-settings==2.13.1

# Logging & Monitoring
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
moto==5.0.2

# Code Quality
black==24.2.0
flake8==7.0.0
mypy==1.8.0

# Utilities
tenacity==8.2.3
tqdm==4.66.1
